requirements/test.txt
-i https://pypi.org/simple
aiohttp==3.6.2
appdirs==1.4.3
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
black==19.3b0
chardet==3.0.4
click==7.0
faker==2.0.3
idna==2.8
importlib-metadata==0.23 ; python_version < '3.8'
more-itertools==7.2.0
multidict==4.5.2
packaging==19.2
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.2
pytest-sanic==1.1.2
pytest==5.2.1
python-dateutil==2.8.0
six==1.12.0
text-unidecode==1.3
toml==0.10.0
wcwidth==0.1.7
yarl==1.3.0
zipp==0.6.0
